+++ qbraid/runtime/native/provider.py
@@ -22,7 +22,7 @@ from qbraid.runtime._display import display_jobs_from_data
 from qbraid.runtime.enums import DeviceActionType, NoiseModel
 from qbraid.runtime.exceptions import ResourceNotFoundError
 from qbraid.runtime.profile import TargetProfile
-from qbraid.runtime.provider import QuantumProvider
+from qbraid.runtime.provider import QuantumProvider, cache_results
 
 from .device import QbraidDevice
 
@@ -95,7 +95,8 @@ class QbraidProvider(QuantumProvider):
             noise_models=noise_models,
         )
 
-    def get_devices(self, **kwargs) -> list[QbraidDevice]:
+    @cache_results(ttl=120)
+    def get_devices(self, bypass_cache: bool = False, **kwargs) -> list[QbraidDevice]:
         """Return a list of devices matching the specified filtering."""
         query = kwargs or {}
         query["vendor"] = "qBraid"
@@ -108,7 +109,12 @@ class QbraidProvider(QuantumProvider):
         profiles = [self._build_runtime_profile(device_data) for device_data in device_data_lst]
         return [QbraidDevice(profile, client=self.client) for profile in profiles]
 
-    def get_device(self, device_id: str) -> QbraidDevice:
+    @cache_results(ttl=120)
+    def get_device(
+        self,
+        device_id: str,
+        bypass_cache: bool = False,
+    ) -> QbraidDevice:
         """Return quantum device corresponding to the specified qBraid device ID.
 
         Returns: