+++ tests/runtime/test_quantum_device.py
@@ -16,6 +16,7 @@ Unit tests for QbraidDevice, QbraidJob, and QbraidJobResult classes using the qb
 """
 import logging
 import random
+import time
 from typing import Any, Optional
 from unittest.mock import Mock, patch
 
@@ -390,6 +391,51 @@ def test_provider_get_devices(mock_client):
     devices = provider.get_devices(qbraid_id="qbraid_qir_simulator")
     assert len(devices) == 1
     assert devices[0].id == "qbraid_qir_simulator"
+
+
+def test_provider_get_cached_devices(mock_client):
+    """Test getting devices from the cache."""
+    client = mock_client
+    provider = QbraidProvider(client=client)
+    client.search_devices = Mock()
+    client.search_devices.return_value = [DEVICE_DATA]
+
+    _ = provider.get_devices(qbraid_id="qbraid_qir_simulator")
+
+    # call again, but ensure that the search method is NOT called again
+    # second call is from cache
+    _ = provider.get_devices(qbraid_id="qbraid_qir_simulator")
+    client.search_devices.assert_called_once()
+
+
+def test_provider_get_devices_post_cache_expiry(mock_client):
+    """Test that the cache entry is invalidated when the cache is too old."""
+    client = mock_client
+    provider = QbraidProvider(client=client)
+    client.search_devices = Mock()
+    client.search_devices.return_value = [DEVICE_DATA]
+
+    init_time = time.time()
+    device_ttl = 120
+    _ = provider.get_devices(qbraid_id="qbraid_qir_simulator")
+
+    # call again, but ensure that the search method is called again and
+    # second call is NOT from cache
+    with patch("time.time", return_value=init_time + device_ttl + 5):
+        _ = provider.get_devices(qbraid_id="qbraid_qir_simulator")
+        assert client.search_devices.call_count == 2
+
+
+def test_provider_get_devices_bypass_cache(mock_client):
+    """Test that the cache is bypassed when the bypass_cache flag is set."""
+    client = mock_client
+    provider = QbraidProvider(client=client)
+    client.search_devices = Mock()
+    client.search_devices.return_value = [DEVICE_DATA]
+
+    _ = provider.get_devices(qbraid_id="qbraid_qir_simulator")
+    _ = provider.get_devices(qbraid_id="qbraid_qir_simulator", bypass_cache=True)
+    assert client.search_devices.call_count == 2
 
 
 def test_provider_search_devices_raises_for_bad_client():